[REFACTOR][DOCS][DEPENDENCY] 엘라스틱 서치 의존성 추가 및 폴더 구조 변경 엘라스틱 서치 관련 설정 yml 변경
