+++ build.gradle
@@ -24,6 +24,7 @@ dependencies {
     runtimeOnly 'com.h2database:h2'
     implementation 'org.springframework.boot:spring-boot-starter-logging'
     implementation 'net.datafaker:datafaker:1.9.0'
+    implementation 'org.springframework.boot:spring-boot-starter-data-elasticsearch'
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
