+++ src/main/java/org/jco/communityservice/exception/CommunityNotFoundException.java
@@ -1,4 +1,4 @@
-package org.jco.communityservice.domain.exception;
+package org.jco.communityservice.exception;
 
 public class CommunityNotFoundException extends RuntimeException{
     public CommunityNotFoundException() {
