+++ src/test/java/org/jco/communityservice/JcoApplicationTests.java
@@ -1,8 +1,7 @@
 package org.jco.communityservice;
 
 import org.jco.communityservice.domain.Category;
-import org.jco.communityservice.domain.Community;
-import org.jco.communityservice.domain.dto.CommunitySaveDto;
+import org.jco.communityservice.dto.CommunitySaveDto;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
